Move MakeWithRandomValues to tests/Utilities
