+++ tests/Unit/CMakeLists.txt
@@ -22,6 +22,7 @@ add_subdirectory(Options)
 add_subdirectory(Parallel)
 add_subdirectory(PointwiseFunctions)
 add_subdirectory(Pypp)
+add_subdirectory(TestUtilities)
 add_subdirectory(Time)
 add_subdirectory(Utilities)
