+++ tests/Unit/Evolution/Systems/RelativisticEuler/Valencia/Test_ConservativeFromPrimitive.cpp
@@ -7,8 +7,8 @@
 #include "DataStructures/Tensor/Tensor.hpp"
 #include "Evolution/Systems/RelativisticEuler/Valencia/ConservativeFromPrimitive.hpp"
 #include "Utilities/MakeWithValue.hpp"
-#include "tests/Unit/DataStructures/TestHelpers.hpp"
 #include "tests/Unit/TestingFramework.hpp"
+#include "tests/Utilities/MakeWithRandomValues.hpp"
 
 namespace {
 template <typename DataType>
