+++ tests/Unit/TestUtilities/CMakeLists.txt
@@ -1,0 +1,15 @@
+# Distributed under the MIT License.
+# See LICENSE.txt for details.
+
+set(LIBRARY "Test_TestUtilities")
+
+set(LIBRARY_SOURCES
+  Test_MakeWithRandomValues.cpp
+  )
+
+add_test_library(
+  ${LIBRARY}
+  "TestUtilities"
+  "${LIBRARY_SOURCES}"
+  "DataStructures"
+  )
